useConfiguration hook
