+++ packages/fe-container/src/App.tsx
@@ -1,28 +1,11 @@
-import React, {useEffect, useState} from 'react'
-import {Configuration} from '@mia-platform/core'
+import React from 'react'
 
 import './App.less'
-import {retrieveConfiguration} from './services/microlc/microlc.service'
 import {Launcher} from './containers/launcher/Launcher'
-import {registerPlugin, finish} from './plugins/PluginsLoaderFacade'
-
-interface AppState {
-  isLoading: boolean,
-  configuration: Configuration
-}
+import {useConfiguration} from './hooks/useConfiguration'
 
 const App: React.FC = () => {
-  const [appState, setAppState] = useState<AppState>({isLoading: true, configuration: {}})
-
-  useEffect(() => {
-    const configurationSubscription = retrieveConfiguration()
-      .subscribe((configuration: Configuration) => {
-        configuration.plugins?.forEach(registerPlugin)
-        finish()
-        setAppState({isLoading: false, configuration})
-      })
-    return () => configurationSubscription.unsubscribe()
-  }, [])
+  const appState = useConfiguration()
 
   return <Launcher {...appState}/>
 }
